Initial implementation of GObject deserialization function
The json_construct_gobject() function takes a GType and a JSON data stream
and constructs a new instance for the given type. If the type is a
JsonSerializable, it will use the JsonSerializable interface for parsing
the JsonNodes of each object member.

This is the initial implementation of the function: the JsonNode-to-GValue
fallback parser is just a stub.