Add the JsonGenerator:root property
JsonGenerator now has a :root property, so it can be constructed using
just g_object_new():

  generator = g_object_new (JSON_TYPE_GENERATOR,
                            "pretty", TRUE,
                            "indent", 4,
                            "root", node,
                            NULL);

This means that the root node is copied inside the generator, instead of
just taking ownership (it was quite confusing). The documentation now
clearly states what happens, and that you can safely free the node after
feeding it to the JsonGenerator.