Change json_parser_get_root() semantics
The json_parser_get_root() returns a pointer to the root node. This does
not conform to the API naming convention inherited from GLib, where
functions returning an internal pointer are called "peek" and function
returning a copy are called "get".

Thus, json_parser_get_root() will now return a copy of the root node and
it is left to the developer to free the returned JsonNode.

A function returning the pointer has also been added, and it's called
json_parser_peek_root().