Copy GScanner into JSON-GLib as JsonScanner
Instead of writing our tokenizer we can fork GScanner and make a
specialized version for JSON (as per RFC), luckily the licenses
are compatible (LGPLv2.1 with "any later" clause).

GScanner does not support Unicode "\uNNNN" escaping and we need to
ensure UTF-8 strings as well.

The API will mostly be the same, but the generic bits not used by
JsonParser will be hidden: this is, after all, a specialized
tokenizer.