Coalesce JsonGenerator test cases into a single test unit
Like we did for JsonParser, JsonGenerator has now a test unit checking the
output of a DOM built and dumped into a buffer.

For the arrays is quite easy to verify that the output is correct;
unfortunately, JsonObject does not guarantee any ordering on the members,
with it being an associative array. Hence, for the object test case we
just compare the length of the output.