Build test-serialize-full
Complete the porting of the old test suite into the new, GTest based one.

The new test suite is automatically run with make test; make test-report
will generate an XML report of the test.

The API coverage for the data types is part of the json-glib/tests
directory and will be expanded later.