Add support for parsing \uXXXX into Unicode
JsonScanner was forked from GScanner in order to support the
JSON-specific Unicode escaping.

JsonScanner now intercepts the \u escape character and, if it
is followed by a hexadecimal value, it will retrieve the
Unicode character encoded in the following 4 values and insert
the character itself in the buffer.

This allows full compatibility with JSON.