Use an array to hold the strings in a vector
Instead of building a GString by concatenating every string inside
an array to deserialize the array into a string vector property,
use a GPtrArray. This is far more efficient (no reallocations are
necessary, as we know the size of the array) and safe (the separator
used to build the string buffer and then split it might exist in
one of the original strings).