Display the filename inside error messages
Instead of just relaying the line number both when parsing files
and memory buffers, JsonParser should also print out the file name
in case it is available.

The error message format should be make-like and emacs-friendly,
that is:

  filename:line_number: error message

so that editors and development environments can parse the errors
easily.

This commit adds the filename string, and a boolean flag for checking
whether the filename is set, inside the JsonParser private data
structure. The boolean flag is checked inside the JsonScanner error
handler when populating the GError or when printing the warning
directly on stderr.