[node] Do not overwrite when copying
Bug 1353 - Copying JSON_NODE_VALUE nodes unreliable at best

When copying a JsonNode to another we do an implicit memcpy using:

  *copy = *src

Which works well enough with pointers, but makes a mess out of the
value-based nodes.

We should just copy the type of the original JsonNode and leave the
rest to the switch() block.

In order to catch potential regressions, we also need a more
thorough test unit for the JsonNode copy operation.