Deprecate add_member() and add set_member() and friends
The add_member() method of JsonObject has wee bit weird semantics: if
the member to be added already exists it prints a scary warning and
returns - and yet it calls g_hash_table_replace() internally as if it
overwrites the member.

So, instead of changing semantics midway we can:

  - add a json_object_set_member() which adds a new member and
    overwrites existing members

  - deprecate json_object_add_member()

While we're at it, we can add convenience wrappers for set_member()
and get_member() that don't require us toying with nodes; luckily,
since the amount of valid types we can add to a JsonObject is limited,
this does not lead to a combinatorial API explosion.