Do not sanitize the object member's name
JsonObject sanitizes the name of the member to replace all
characters defined by G_STR_DELIMITERS with '_'. This is
absolutely brain damaged, since a member name can be any
valid JSON string.

Obviously, if a member name maps to a GObject property is
entirely up to the GObject code to decide whether to sanitize
the member name or not.