object: Guarantee insertion order
When iterating over the members of a JsonObject, or when retrieving
the list of members, the insertion order should be preserved by the
JsonObject. This is simply implemented by keeping a mirror list of
the member names.

Apparently, though JSON does not guarantee any ordering, it is somewhat
expected by JSON (and ECMAScript) users.