build: Clean up the build environment
Some of the rules can be moved into their own files to be included
when needed, like:

  • silent rules (QUIET_*)
  • glib-mkenums rules
  • glib-genmarshal rules

Also, the test suite rules should be moved from the top-level of
the project into the build/autotools directory and then included
only where it makes sense.

This requires changing most of the build system to use the new
files layout.