gobject: Support constructor properties
The GObject deserialization code currently skips all the constructor
and constructor-only properties. In order to implement them we can
add a preliminary pass on the JSON object members and build a
GParameter array.

As we don't have a GObject instance we cannot really use the
Serializable interface to provide custom parsing for complex data
structures, thus we fall back to the default deserialization code
path.