object: Return values list in insertion order
Since we return the member names in insertion order, we should also
return the member values in the same order.

This also allows us to get rid of the (yucky) internal copies of
g_hash_table_get_keys() and g_hash_table_get_values(), since we use
the hash table only for storage and lookup purposes.