boxed: Split (de)serialization registration
A GBoxed type defined as:

  struct Boxed {
    int foo;
    gboolean bar;
    int baz;
  };

Can be represented either by a JSON object:

  {
    "foo" : 1,
    "bar" : true,
    "baz" : 3
  }

Or by a JSON array:

  [ 1, true, 3 ]

The current function for registering a serialization and a
deserialization pair does not allow registering more than one
deserialization function - which means that there can only be
one way to deserialize a GBoxed type into a specific JsonNode
type.

To allow having more than one JsonNodeType associated to a
GBoxed type and a deserialization function we need to split out
the registration of the serialization and deserialization functions
into two distinct functions.