Add GCOV coverage test support
Use GCOV, a GCC extension, to perform a coverage test on JSON-GLib when
we perform a 'make check'.

GCOV support builds JSON-GLib with compiler and linker flags that enable
coverage reports; then the check-local target will build the gcov file
for each source file, and the coverage report will be extracted from
those.