tests: Move Parser and Generator tests
Use the json-glib/tests directory for testing the data structures of
JSON-GLib: node, object, array, parser and generator.

The tests/ directory should be used for complex test cases, like the
GObject and GBoxed integration - but the goal is to remove the top-level
tests/ directory altogether, since the conformance test suite should be
built along the json-glib/ directory.