parser: Refactor the JsonParser logic
The array and object parsing logic in JsonParser has clearly exploded
beyond control: a simple tightening of the JSON validation almost broke
the parser in two. It it is time to...

        <cue Christopher Lee voice-over>

        REFACTOR THE CODE!

        </cue Christopher Lee voice-over>

This time, we should be following the JSON state machine and try to do
more prediction of the next state based on peeking the next token.

The code is fairly cleaner, now; and, most of all, still passes the
validation test suite - which is, you know... nice.

(cherry picked from commit 08cec3b7c9202007a5abbf548f8d2f1b54d4d0b4)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>