reader: Allow using read_element() on objects
If we assume that a JSON object is just an array with a named mapping
then the JsonReader API should be able to descend into objects using the
same API used for arrays.

This obviously is less useful than it sounds if we take a very strict
interpretation of JSON objects as unordered string-to-value mappings; as
the ordering is not guaranteed to be stable, parsers would be fairly
weak against various JSON definitions.

If the JSON format parsed is guaranteed to be stable then an integer
offset might be an easy (albeit slightly less performant) way to access
data.