object: Replace the name pointer in the members list
When calling g_hash_table_replace() we also free the string holding the
member name. This means that the const gchar* pointer we store inside
the list of ordered member names now points to garbage - so if somebody
tries to iterate over the members list it will get random values instead
of a valid C string.

Since we guaranteed insertion order, if we replace the contents of a
JsonObject member we need to find the right pointer and replace it: just
removing and prepending won't do.

https://bugzilla.gnome.org/show_bug.cgi?id=642383
(cherry picked from commit 1a633159a593c962233a5ef4660e31e60eed96d9)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>