build: Add abicheck.sh
The abicheck.sh script is used to check that the actural ABI we expose
in our shared library is the same as the expected one we encode inside
the json-glib.symbols file.

Using abicheck.sh at check time, we can verify that: a) the symbols file
is up to date and b) we're not unintentionally leaking private symbols
to the outside world.