build: Drop unnecessary linker flags
The -no-undefined linker flag is needed only when compiling for Windows,
and it's added conditionally. Except that we also add it unconditionally
in the LDFLAGS used for the libjson-glib-1.0 shared object. Whoopsie.

We should also drop the unneeded -rpath: distributions will go out of
their way to patch it out in any case.