Overhaul the Visual Studio 2010 Projects
Split up the property sheets used to build JSON-GLib. Also add projects to
build and "install" the json-glib-validate and json-glib-format tools.

Move the introspection file generation out of the Visual Studio projects
to add an NMake Makefile later to generate the introspection files, to make
the process clearer to people.

Stop installing the test programs, although they continue to be built from
the projects for now.