MSVC Builds: Rework Introspection Files Build
Use a set of NMake Makefiles in place of custom Visual Studio Projects to
generate the introspection files for JSON-GLib.

Although this means that one needs to build the introspection files in two
stages, namely using the projects to build the JSON-GLib DLL and then
running this NMake Makefile to generate the introspection files.  The
original approach required setting lots of environment variables for the
building of introspection files to work.  Using NMake Makefiles will make
it clearer to the one building the introspection files on what items and
variables are needed exactly to do the job