Visual Studio 2010 Projects: Add PlatformToolset Tag
This adds a PlatformToolset tag to each configuration so that the projects
can be updated to Visual Studio 2012 or 2013 formats quite easily, using
Visual Studio's round-tripping feature which actually looks at this tag.

This, for example, can be quite easily processed by a script to produce
the Visual Studio 2012 and 2013 projects.