MSVC Builds: Speed Up Release Builds
Use the multiprocessor compilation (/MP) option, which can cut down release
build times by quite a bit.  This will emit a brief warning on debug
builds, but otherwise the build will proceed as normal.

Also use the /d2Zi+ option for MSVC 2010 (and later) builds so that more
useful information can be logged to the .pdb files for the release builds.