Don't loose decimal in whole-double -> string conversion
When converting json to its string representation, whole-doubles (such
as 1.0) would be converted into strings without decimals ("1"). That can
be inconvenient e.g. when converting from/to GVariants.

To avoid this, append '.0' to the string representation for doubles if
they lost their decimals in the conversion.

Also add / update unit tests for this.

https://bugzilla.gnome.org/show_bug.cgi?id=753763