node: Add json_node_ref() and json_node_unref()
Add reference counting semantics to JsonNode, in addition to the
existing init/unset and alloc/free semantics.

json_node_free() must only be used with nodes allocated using
json_node_alloc(). json_node_unref() may be used with all nodes (if
correctly paired; it may be paired with json_node_alloc()).

It is not valid to call json_node_free() on a node whose reference count
is not 1.

https://bugzilla.gnome.org/show_bug.cgi?id=756121