build: Clean up MSVC-related items in autotools build files
This makes use of the common build/Makefile.msvcproj to create the full
Visual Studio 2008/2010 project files from their respective templates,
while making the autotools files cleaner.  As an added benefit, this also
enables the list of headers to "install" to be filled in automatically
during 'make dist', so that we will have to worry less about the addition
or removal of public headers.