Visual Studio builds: Support Visual Studio 2017
Support Visual Studio 2017 by updating the scripts that copy the 2010
projects and updates items as neccessary to produce the 2017 projects.

Please note that since the toolset version string for Visual Studio 2017
changed, a custom toolset version string is allowed and used if specified,
otherwise a toolset version string is generated as it were before.

Note also that Visual Studio 2017 aims to be compatible with 2015 on the
CRT level, so there should not be problems using 2017-built binaries with
2015-built binaries.