build: Fix soversion for Meson builds
During stable cycles we want the interface age to match the micro
version, as there are no symbol additions if the minor version is
an even number.

Conversely, we want the interface age to reset to 0, as there are
no expectations of ABI compatibility for newly added symbols.

The current check uses the micro version, instead of the odd version,
which breaks the logic above.

This ensures that the soname generated via the Meson build is the
same as the one generated by the Autotools one.